Fitur Admin

+++ app/Http/Controllers/ItemsController.php
@@ -2,8 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Aucation;
 use App\Models\Item;
+use App\Models\Brand;
 use App\Models\ItemImage;
+use App\Models\ItemCategory;
+use App\Models\ItemDetail;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
@@ -19,7 +23,12 @@ class ItemsController extends Controller
     {
         $data['page_title'] = "List Barang";
         $data['title'] = "Barang";
-        $data['items'] = Item::all();
+        if (Auth()->guard('officer')->user()->level_id == 1) {
+            $data['items'] = Item::all();
+            # code...
+        }else{
+            $data['items'] = Item::where('officer_id', Auth()->guard('officer')->user()->officer_id)->get();
+        }
 
         return view('Admin.DataTables.DataBarang', $data);
     }
@@ -33,6 +42,9 @@ class ItemsController extends Controller
     {
         $data['page_title'] = "Tambah Barang";
         $data['title'] = "Barang";
+        $data['categories'] = ItemCategory::all();
+        $data['brands']     = Brand::all();
+
         return view('Admin.DataTables.TambahBarang', $data);
     }
 
@@ -42,34 +54,61 @@ class ItemsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
+
+    public function itemReqValidate($request){
         $cusMessage = [
             "item_name.required" => "Nama barang harus di isi",
-            "initial_price.required" => "Harga awal harus di isi",
             "images.required" => "Gambar harus di isi",
+            "camera.required" => "Resolusi kamera harus di isi",
+            "display"         => "Resolusi layar harus di isi",
+            "battery.required" => "Kapasitas batrai harus di isi",
+            "chipset.required" => "Chipset barang harus di isi",
+            "storage.required" => "Kapasitas penyimpanan barang harus di isi",
             "images.*.mimes" => "Format gambar harus berupa jpg, png, jpeg atau gif",
-            "images.*.max" => "Ukuran maksimal gambar 2MB",
+            "images.*.max" => "Ukuran maksimal gambar 2MB"
         ];
         // dd($request->images);
         $request->validate([
             'item_name' => 'required',
-            'initial_price' => 'required',
             'description'   => 'required',
+            'display'   => 'required',
+            'camera'   => 'required',
+            'chipset'   => 'required',
+            'battery'   => 'required',
+            'storage'   => 'required',
             'images'  => 'required',
             'images.*' => 'mimes:jpg,jpeg,png,gif|max:2000'
         ], $cusMessage);
+    }
 
+    public function storeDetail($data, $param){
+        $storeDetail = ItemDetail::create([
+            "camera" => $data->camera,
+            "display"=> $data->display,
+            "chipset"=> $data->chipset,
+            "battery"=> $data->battery,
+            "storage"=> $data->storage,
+            "brand_id" => $data->brand,
+            "condition"=>$data->condition,
+            "item_id" => $param
+        ]);
+    }
+
+    public function store(Request $request)
+    {
+        
+        $this->itemReqValidate($request);
         
         DB::transaction(function() use($request){
             $storeItems = Item::create([
                 "item_name" => $request->item_name,
-                "initial_price" => $request->initial_price,
                 "description" => $request->description,
-                "category_id" => 1,
-                "officer_id"  => 1,
+                "category_id" => $request->category,
+                "officer_id"  => Auth()->guard('officer')->user()->officer_id,
                 "item_main_image" => $request->file('images')[0]->store('itemImages')
             ]);
+
+            $this->storeDetail($request, $storeItems->item_id);
 
             foreach($request->file('images') as $image){
                 if($image->getClientOriginalName() == $request->file('images')[0]->getClientOriginalName()){
@@ -84,7 +123,7 @@ class ItemsController extends Controller
                 ]);
             }
         });
-        return redirect()->route('listBarang')->with('message', "Data berhasil ditambahkan");
+        return redirect()->route('listBarang')->with('success', "Data berhasil ditambahkan");
     }
 
     /**
@@ -114,6 +153,9 @@ class ItemsController extends Controller
         $data['page_title'] = "Edit Barang";
         $data['title'] = "Barang";
         $data['item'] = $item;
+        $data['categories'] = ItemCategory::all();
+        $data['brands']     = Brand::all();
+
         return view('Admin.DataTables.TambahBarang', $data);
     }
 
@@ -124,6 +166,18 @@ class ItemsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+
+    public function updateDetail($data, $param){
+        $updateDetail = ItemDetail::where('item_id', $param)->update([
+            "camera" => $data->camera,
+            "display"=> $data->display,
+            "chipset"=> $data->chipset,
+            "battery"=> $data->battery,
+            "storage"=> $data->storage,
+            "brand_id" => $data->brand,
+            "condition"=>$data->condition,
+        ]);
+    }
     public function update(Request $request, Item $item)
     {
         // foreach($request->imagesDelete as $image){
@@ -132,18 +186,24 @@ class ItemsController extends Controller
         // dd($request->all());
         $cusMessage = [
             "item_name.required" => "Nama barang harus di isi",
-            "initial_price.required" => "Harga awal harus di isi",
-            "image.required" => "Gambar harus di isi",
+            "images.required" => "Gambar harus di isi",
+            "camera.required" => "Resolusi kamera harus di isi",
+            "display"         => "Resolusi layar harus di isi",
+            "battery.required" => "Kapasitas batrai harus di isi",
+            "chipset.required" => "Chipset barang harus di isi",
+            "storage.required" => "Kapasitas penyimpanan barang harus di isi",
             "images.*.mimes" => "Format gambar harus berupa jpg, png, jpeg atau gif",
-            "images.*.max" => "Ukuran maksimal gambar 2MB",
-            "image.mimes" => "Format gambar harus berupa jpg, png, jpeg atau gif",
-            "image.max" => "Ukuran maksimal gambar 2MB",
+            "images.*.max" => "Ukuran maksimal gambar 2MB"
         ];
         // dd($request->images);
         $request->validate([
             'item_name' => 'required',
-            'initial_price' => 'required',
             'description'   => 'required',
+            'display'   => 'required',
+            'camera'   => 'required',
+            'chipset'   => 'required',
+            'battery'   => 'required',
+            'storage'   => 'required',
         ], $cusMessage);
 
         if($request->hasFile('image')){
@@ -151,14 +211,22 @@ class ItemsController extends Controller
                 'image' => 'required|mimes:jpg,jpeg,png,gif|max:2048'
             ], $cusMessage);
             $item->item_main_image !== null ? Storage::delete($item->item_main_image) : '';
-            $this->queryUpdateImages($request, $item);
-            
+            DB::transaction(function() use($request, $item){
+                $this->queryUpdateImages($request, $item);
+                Item::where('item_id', $item->item_id)->update([
+                    "item_name" => $request->item_name,
+                    "description" => $request->description,
+                ]);
+                $this->updateDetail($request, $item->item_id);
+            });
         }else{
-            Item::where('item_id', $item->item_id)->update([
-                "item_name" => $request->item_name,
-                "initial_price" => $request->initial_price,
-                "description" => $request->description,
-            ]);
+            DB::transaction(function() use($request, $item){
+                Item::where('item_id', $item->item_id)->update([
+                    "item_name" => $request->item_name,
+                    "description" => $request->description,
+                ]);
+                $this->updateDetail($request, $item->item_id);
+            });
         }
         if(isset($request->imagesDelete) && $request->file('images')){
             $request->validate([
@@ -200,13 +268,12 @@ class ItemsController extends Controller
             }
         }
 
-        return redirect()->route('listBarang')->with('message', 'Data barang berhasil diubah');
+        return redirect()->route('listBarang')->with('success', 'Data barang berhasil diubah');
     }
 
     public function queryUpdateImages($data, $params){
         Item::where('item_id', $params->item_id)->update([
             "item_name" => $data->item_name,
-            "initial_price" => $data->initial_price,
             "description" => $data->description,
             "item_main_image" => $data->image->store('itemImages')
         ]);
@@ -237,6 +304,10 @@ class ItemsController extends Controller
     public function destroy(Item $item)
     {
         $parent = Item::find($item->item_id);
+        if(Aucation::where('item_id', $item->item_id)->get()){
+            return back()->with('error', 'Barang ini sedang di lelang');
+        }
+
         foreach ($parent->images as $value) {
             if($value){
                 Storage::delete($value->image_path);
@@ -246,6 +317,6 @@ class ItemsController extends Controller
         Storage::delete($parent->item_main_image);
         $parent->delete();
 
-        return back();
+        return back()->with('success', 'Berhasil menghapus data barang');
     }
 }
